Add prime numbers service call to controller
The PrimeNumberController call a client factory that returns a
client/connection to the PrimeNumbersService. This is used to call the
primeNumbersUtil by doing a rpc.

Also, two callbacks were setup for both success and failure. In case of
success, the endpoint should return a 200 with the prime numbers as a
json array. In case of failure, a ServiceFailureException is raised and
the endpoint returns a 500.