Add dependency injection and feature tests
Added the dependency injection to the project so that the ClientsFactory
could be passed in the controllers in a easy way. This meant that I had
to add ClientsFactoryModule that is the helper class that allows Finatra
to take care of the injections.

This also helped with the mocking of the PrimeNumbersService in the
feature tests.